requirements.txt
Django==5.0 
djangorestframework==3.14.0 
django-storages==1.14.2
mysqlclient==2.2.1
pillow==10.2.0